Refactor: rename list_estimators.py to query_registry.py
