+++ src/sktime_mcp/tools/query_registry.py
